[KFI]refactor(document-viewer): move create-comment component higher in tree
