+++ packages/sn-document-viewer-react/src/components/document-widgets/index.ts
@@ -14,4 +14,3 @@ export { ToggleWatermarkWidget, ToggleWatermarkComponent } from './ToggleWaterma
 export { ToggleCommentsWidget, ToggleComments } from './ToggleComments'
 export { ZoomInOutWidget } from './ZoomInOut'
 export { ZoomModeWidget } from './ZoomMode'
-export { CreateComment } from './CreateComment'
