+++ packages/sn-document-viewer-react/src/components/index.ts
@@ -9,3 +9,4 @@ export * from './document-widgets'
 export * from './page-widgets'
 export * from './styles'
 export { CommentComponent } from './Comment'
+export { CreateComment } from './CreateComment'
